requirements.txt
discord.py==2.3.2
yt-dlp==2024.4.9
python-dotenv==1.0.1